windows

+++ tests/integration/free_vs_w_parent_test.py
@@ -9,11 +9,11 @@ def test_free_vs_w_parent(library: str) -> None:
     df1 = integer_dataframe_1(library)
     namespace = df1.__dataframe_namespace__()
     free_ser1 = namespace.column_from_1d_array(  # type: ignore[call-arg]
-        np.array([1, 2, 3]),
+        np.array([1, 2, 3], dtype="int64"),
         name="preds",
     )
     free_ser2 = namespace.column_from_1d_array(  # type: ignore[call-arg]
-        np.array([4, 5, 6]),
+        np.array([4, 5, 6], dtype="int64"),
         name="preds",
     )
